### 2019-07-29 #1607 @whedon
[PRE REVIEW]: Ryacas: A computer algebra system in R
**Submitting author:** @mikldk (<a href="http://orcid.org/0000-0002-0234-0266">Mikkel Meyer Andersen</a>)
**Repository:** <a href="https://github.com/mikldk/ryacas" target ="_blank">https://github.com/mikldk/ryacas</a>
**Version:** v1.0.0
