### 2019-08-01 #1607 @labarba
@mikldk —
After a pre-review of your submission, and an assessment by the editorial board at large, we find that this submission does not quite meet our [submission requirements](http://joss.theoj.org/about#submission_requirements). In particular, while the software seems very useful, it falls under the 'minor utility/thin wrapper' category, and we therefore decided not to put it through review.

Thank you for considering JOSS as a venue for your software, and I hope you'll submit other software with research application in the future.
