### 2019-08-02 7ff489a
Remove paper
