+++ DESCRIPTION
@@ -1,5 +1,5 @@
 Package: Ryacas
-Version: 0.9.0.9030
+Version: 0.9.0.9031
 Title: R Interface to the Yacas Computer Algebra System
 Authors@R: c(
     person(given = "Mikkel Meyer", family = "Andersen",
@@ -38,5 +38,5 @@ SystemRequirements: C++11
 URL: https://github.com/mikldk/ryacas, http://www.yacas.org
 BugReports: https://github.com/mikldk/ryacas/issues
 Roxygen: list(markdown = TRUE, roclets = c("rd", "collate", "namespace"))
-RoxygenNote: 6.1.1
+RoxygenNote: 6.1.1.9000
 VignetteBuilder: knitr
