+++ R/yac-symbol.R
@@ -1,0 +1,615 @@
+# Supported functions for yac_symbol()
+#######################################
+#
+# ------------------------------------
+# From Ryacas:
+# ------------------------------------
+# * yac_str()
+# * yac_expr()
+# * yac_silent()
+# * yac()
+# * yac_assign()
+#
+# New:
+# * simplify()
+# * tex()
+# 
+# ------------------------------------
+# From R:
+# ------------------------------------
+# * print()
+# * dim()
+# * `[` getter
+# * `[<-` setter
+# * %*% 
+# * diag(): both getter and setter
+# * upper.tri(): only getter
+# * lower.tri(): only getter
+# * t()
+# * solve()
+# * Ops: c("+", "-", "*", "/", "^")
+# * Math functions: 
+#    sin(), cos(), tan(), 
+#    asin(), acos(), atan(), asinh(), acosh(), atanh(), 
+#    exp(), log(), sqrt()
+# 
+# ------------------------------------
+# Missing:
+# ------------------------------------
+# * upper.tri(): setter
+# * lower.tri(): setter
+# 
+# ------------------------------------
+# Not included on purpose:
+# ------------------------------------
+# * 
+
+
+
+
+#' Make a yacas symbol
+#' 
+#' Note that this results in multiple calls to `yacas` and 
+#' the performance may be slower than manually using e.g. [yac_str()].
+#' 
+#' @param x A vector or a matrix
+#' 
+#' @return A `yac_symbol`
+#' 
+#' @concept yac_symbol
+#' @export
+yac_symbol <- function(x) {
+  stopifnot(is.vector(x) | is.matrix(x))
+  
+  cmd <- if (is.matrix(x) | length(x) > 1L) {
+    as_y(x)
+  } else {
+    x
+  }
+  
+  is_mat <- grepl("^[ ]*\\{[ ]*\\{.*\\}[ ]*\\}[ ]*$", cmd)
+  is_vec <- !is_mat && grepl("^[ ]*\\{.*\\}[ ]*$", cmd)
+  
+  y <- list(yacas_cmd = cmd,
+            is_mat = is_mat,
+            is_vec = is_vec)
+  
+  class(y) <- "yac_symbol"
+  return(y)
+}
+
+# S3 exports
+
+#' @export
+yac_str.yac_symbol <- function(x) {
+  return(yac_str(x$yacas_cmd))
+}
+
+#' @export
+yac_expr.yac_symbol <- function(x) {
+  return(yac_expr(x$yacas_cmd))
+}
+
+#' @export
+yac_silent.yac_symbol <- function(x) {
+  return(yac_silent(x$yacas_cmd))
+}
+
+#' @export
+yac.yac_symbol <- function(x, rettype = c("str", "expr", "silent")) {
+  return(yac(x$yacas_cmd, rettype))
+}
+
+#' @export
+yac_assign.yac_symbol <- function(value, x) {
+  return(yac_silent(paste0(x, " := ", value$yacas_cmd)))
+}
+
+#' @export
+y_fn.yac_symbol <- function(x, fn, ...) {
+  x <- y_fn(x$yacas_cmd, fn, ...)
+  y <- yac_str(x)
+  z <- yac_symbol(y)
+  return(z)
+}
+
+
+expr_has_vars <- function(x) {
+  y_vars <- all.vars(x)
+  
+  if (length(y_vars) > 0L) {
+    return(TRUE)
+  }
+  
+  return(FALSE)
+}
+
+#' @export
[…]
