+++ R/legacy.R
@@ -1,138 +1,3 @@
-# if (FALSE) {
-#   exports <- '
-# S3method("%Where%",default)
-# S3method(Clear,Expr)
-#   S3method(Clear,default)
-#   S3method(Conjugate,Expr)
-#   S3method(Conjugate,default)
-#   S3method(Eval,Expr)
-#   S3method(Eval,Sym)
-#   S3method(Eval,default)
-#   S3method(Eval,yacas)
-#   S3method(Expand,Expr)
-#   S3method(Expand,default)
-#   S3method(Factor,Expr)
-#   S3method(Factor,default)
-#   S3method(Factorial,Expr)
-#   S3method(Factorial,default)
-#   S3method(Identity,default)
-#   S3method(Integrate,Expr)
-#   S3method(Integrate,default)
-#   S3method(Inverse,Expr)
-#   S3method(Inverse,default)
-#   S3method(InverseTaylor,default)
-#   S3method(Limit,Expr)
-#   S3method(Limit,default)
-#   S3method(List,Expr)
-#   S3method(List,default)
-#   S3method(Math,Expr)
-#   S3method(Math,Sym)
-#   S3method(N,Expr)
-#   S3method(N,default)
-#   S3method(Newton,Expr)
-#   S3method(Newton,default)
-#   S3method(Ops,Expr)
-#   S3method(Ops,Sym)
-#   S3method(Precision,Expr)
-#   S3method(Precision,default)
-#   S3method(PrettyForm,Expr)
-#   S3method(PrettyForm,default)
-#   S3method(PrettyPrinter,Expr)
-#   S3method(PrettyPrinter,default)
-#   S3method(Simplify,Expr)
-#   S3method(Simplify,default)
-#   S3method(Solve,Expr)
-#   S3method(Solve,default)
-#   S3method(Subst,default)
-#   S3method(Taylor,Expr)
-#   S3method(Taylor,default)
-#   S3method(TeXForm,default)
-#   S3method(Transpose,default)
-#   S3method(Ver,Expr)
-#   S3method(Ver,default)
-#   S3method(as.Sym,Expr)
-#   S3method(as.Sym,character)
-#   S3method(as.Sym,matrix)
-#   S3method(as.Sym,yacas)
-#   S3method(as.character,Expr)
-#   S3method(as.character,Sym)
-#   S3method(as.character,yacas)
-#   S3method(as.expression,Sym)
-#   S3method(as.expression,yacas)
-#   S3method(deriv,Expr)
-#   S3method(deriv,Sym)
-#   S3method(determinant,Expr)
-#   S3method(determinant,Sym)
-#   S3method(print,Expr)
-#   S3method(print,Sym)
-#   S3method(print,yacas)
-#   S3method(yacas,"function")
-#   S3method(yacas,character)
-#   S3method(yacas,expression)
-#   S3method(yacas,formula)
-#   S3method(yacas,yacas)
-#   export("%Where%")
-#   export(Clear)
-#   export(Conjugate)
-#   export(Eval)
-#   export(Expand)
-#   export(Expr)
-#   export(Exprq)
-#   export(Factor)
-#   export(Factorial)
-#   export(I)
-#   export(Identity)
-#   export(Infinity)
-#   export(Integrate)
-#   export(Inverse)
-#   export(InverseTaylor)
-#   export(Limit)
-#   export(List)
-#   export(N)
-#   export(Newton)
-#   export(Pi)
-#   export(Precision)
-#   export(PrettyForm)
-#   export(PrettyPrinter)
-#   export(Ryacas_options)
-#   export(Set)
-#   export(Simplify)
[…]
