+++ man/Math.yac_symbol.Rd
@@ -1,0 +1,14 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/yac-symbol.R
+\name{Math.yac_symbol}
+\alias{Math.yac_symbol}
+\title{Math functions}
+\usage{
+\method{Math}{yac_symbol}(x, ...)
+}
+\arguments{
+\item{x}{A \code{yac_symbol}.}
+}
+\description{
+Math functions
+}

[…]
