### 2019-08-06 767bb6d
don't force unix suggest on AppVeyor
