+++ vignettes/arbitrary-precision.Rmd
@@ -1,0 +1,167 @@
+---
+title: "Arbitrary-precision arithmetic"
+author: "Mikkel Meyer Andersen and Søren Højsgaard"
+date: "`r Sys.Date()`"
+output: rmarkdown::html_vignette
+vignette: >
+  %\VignetteIndexEntry{Arbitrary-precision arithmetic}
+  %\VignetteEngine{knitr::rmarkdown}
+  %\VignetteEncoding{UTF-8}
+---
+
+```{r setup, include=FALSE}
+knitr::opts_chunk$set(echo = TRUE)
+```
+
+```{r, message=FALSE}
+library(Ryacas)
+```
+
+In `R`, the package `Rmpfr` package provide some functions for arbitrary-precision arithmetic. 
+Below we will revisit some of `Rmpfr`'s examples done with `Ryacas` as 
+well at demonstrate new ones possible with `Ryacas` but not with `Rmpfr` (e.g. solving systems of linear equations and matrix inverses).
+
+# Arbitrary-precision arithmetic
+
+First see the problem when using floating-point arithmetic:
+
+```{r}
+(0.3 - 0.2) - 0.1
+yac_str("(0.3 - 0.2) - 0.1") # decimal number does not 
+                             # always work well; often 
+                             # it is better to represent as 
+                             # rational number if possible
+
+# (1/3 - 1/5) = 5/15 - 3/15 = 2/15
+(1/3 - 1/5) - 2/15
+yac_str("(1/3 - 1/5) - 2/15")
+```
+
+
+The `yacas` function [`N()`](https://yacas.readthedocs.io/en/latest/reference_manual/arithmetic.html#N) gives the numeric (floating-point) value for a precision.
+
+```{r}
+yac_str("1/3")
+yac_str("N(1/3)")
+yac_str("N(1/3, 1)")
+yac_str("N(1/3, 200)")
+```
+
+
+# Evaluating polynomials
+
+Consider the polynomial $(x-1)^7$ with root 1 (with multiplicity 7). We can consider it both in factorised form and in expanded form:
+```{r}
+pol <- "(x-1)^7"
+p1 <- pol %>% yac_expr()
+p1
+p2 <- pol %>% y_fn("Expand") %>% yac_expr()
+p2
+```
+
+Mathematically, these are the same objects, but when it comes to numerical
+computations, the picture is different:
+
+```{r}
+eval(p1, list(x = 1.001))
+eval(p2, list(x = 1.001))
+```
+
+We can of course also evaluate the polynomials in the different forms in `yacas` using [`WithValue()`](https://yacas.readthedocs.io/en/latest/reference_manual/controlflow.html#WithValue):
+
+```{r}
+# First try with 1.001:
+pol_val <- paste0("WithValue(x, 1.001, ", pol, ")")
+pol_val
+yac_str(pol_val)
+#... to get result symbolically, use instead the number as a fraction
+pol_val <- paste0("WithValue(x, 1001/1000, ", pol, ")")
+pol_val
+yac_str(pol_val)
+pol_val %>% y_fn("Denom") %>% yac_str()
+pol_val %>% y_fn("Denom") %>% y_fn("IntLog", "10") %>% yac_str()
+pol_val <- paste0("WithValue(x, 1001/1000, Expand(", pol, "))")
+pol_val
+yac_str(pol_val)
+```
+
+The reason for the difference is the near cancellation problem:
+Subtraction of nearly identical quantities. This phenomenon is
+illustrated in the plot below.
+
[…]
