### 2019-08-07 b02400d
work on high-level, tests and vignettes
