+++ vignettes/getting-started.Rmd
@@ -50,6 +50,15 @@ The high-level is easier dealing with vectors and matrices, but may also be less
 Below, we will demonstrate both interfaces and refer to the other vignettes 
 for more information.
 
+A short summary of often-used `yacas` commands are found 
+at the end of this vignette.
+A short summary of often-used low-level `Ryacas` functions are found 
+at the end of the [The low-level interface](./low-level.html) vignette, 
+and a short summary of often-used high-level `Ryacas` functions are found 
+at the end of the [The high-level interface](./high-level.html) vignette.
+
+
+
 ## The low-level interface
 
 The low-level interface consists of these two main functions:
@@ -59,16 +68,16 @@ The low-level interface consists of these two main functions:
 
 Note, that the `yacas` command `x` is a string and must often be built op using `paste()`/`paste0()`. 
 Examples of this will be shown in multiple examples below. 
-Short summaries of often used `yacas` commands and `Ryacas` functions are found 
-at the end of the [The low-level interface](./low-level.html) vignette.
 
-## Examples
+
+### Examples
 
 ```{r}
 eq <- "x^2 + 4 + 2*x + 2*x"
 yac_str(eq) # No task was given to yacas, so we simply get the same returned
 yac_str(paste0("Simplify(", eq, ")"))
 yac_str(paste0("Factor(", eq, ")"))
+yac_expr(paste0("Factor(", eq, ")"))
 yac_str(paste0("TeXForm(Factor(", eq, "))"))
 ```
 
@@ -81,7 +90,7 @@ yac_str(y_fn(eq, "Factor"))
 yac_str(y_fn(y_fn(eq, "Factor"), "TeXForm"))
 ```
 
-As you see, there are a lot of nested function call. That can be avoided by using `magrittr`'s pipe `%>%` (automatically available with `Ryacas`) together with the helper function `y_fn()`:
+As you see, there are a lot of nested function calls. That can be avoided by using `magrittr`'s pipe `%>%` (automatically available with `Ryacas`) together with the helper function `y_fn()`:
 
 ```{r}
 eq %>% y_fn("Factor")
@@ -89,7 +98,7 @@ eq %>% y_fn("Factor") %>% yac_str()
 eq %>% y_fn("Factor") %>% y_fn("TeXForm") %>% yac_str()
 ```
 
-The polynomial can be evaluated for a value of $x$ by calling [yac_expr()] instead of [yac_str()]:
+The polynomial can be evaluated for a value of $x$ by calling `yac_expr()` instead of `yac_str()`:
 ```{r}
 yac_str(paste0("Factor(", eq, ")"))
 expr <- yac_expr(paste0("Factor(", eq, ")"))
@@ -98,331 +107,84 @@ eval(expr, list(x = 2))
 ```
 
 
-### The high-level interface
 
-The high-level interface consists of the main function `yac_symbol()`.
+## The high-level interface
 
-# Examples of 
+The high-level interface consists of the main function `yac_symbol()` and 
+often the helper function `as_r()` will be used to get back an `R` object (expression, matrix, vector, ...).
 
-## Example 1: Simple algebra
 
-First, consider this polynomial:
+### Examples
 
+Before we had `eq` as a text string. We now make a `yac_symbol()` from that:
 ```{r}
[…]
