### 2019-08-07 8a4b0b2
Polishing and building vignettes, docs etc
