+++ man/upper.tri.Rd
@@ -9,6 +9,8 @@ upper.tri(x, diag = FALSE)
 \arguments{
 \item{x}{If \code{yac_symbol} treat as such, else
 call \code{\link[base:lower.tri]{base::lower.tri()}}/\code{\link[base:upper.tri]{base::upper.tri()}}.}
+
+\item{diag}{Whether diagonal is included.}
 }
 \description{
 Lower and upper triangular part of a matrix
