+++ NAMESPACE
@@ -21,6 +21,7 @@ S3method(lower.tri,yac_symbol)
 S3method(print,yac_symbol)
 S3method(simplify,yac_symbol)
 S3method(solve,yac_symbol)
+S3method(str,yac_symbol)
 S3method(t,yac_symbol)
 S3method(tex,yac_symbol)
 S3method(upper.tri,default)
@@ -107,6 +108,7 @@ export(yacas_evaluate)
 export(yacmode)
 importFrom(Rcpp,evalCpp)
 importFrom(magrittr,"%>%")
+importFrom(methods,is)
 importFrom(stats,deriv)
 importFrom(utils,loadhistory)
 importFrom(utils,savehistory)
