+++ DESCRIPTION
@@ -22,6 +22,7 @@ Depends: R (>= 3.3.0)
 Imports:
     Rcpp (>= 0.12.0),
     stats,
+    methods,
     magrittr
 LinkingTo: Rcpp
 Suggests:
@@ -33,7 +34,8 @@ Suggests:
     rmarkdown,
     igraph,
     testthat,
-    unix
+    unix,
+    Rmpfr
 License: GPL
 SystemRequirements: C++11
 URL: https://github.com/mikldk/ryacas, http://www.yacas.org
