+++ README.Rmd
@@ -46,11 +46,14 @@ devtools::install_github("mikldk/ryacas")
 For vignettes, overview, pointers to additional information, installation
 instructions and a sample session see <http://mikldk.github.io/ryacas/>. 
 
-Yacas documentation can be found at http://yacas.readthedocs.org/
+Yacas documentation can be found at http://yacas.readthedocs.org/.
 
 ## Brief examples ##
 
-Below we show a few examples. Please refer to the ["Getting started"](http://mikldk.github.io/ryacas/articles/getting-started.html) vignette (and the other vignettes) for more information.
+Below we show a few examples. 
+We highly recommend reading the "[Getting started](http://mikldk.github.io/ryacas/articles/getting-started.html)" vignette (and the other vignettes) for a more thorough introduction to the package.
+
+### Low-level interface
 
 Returning strings with `yac_str()`:
 
@@ -69,7 +72,7 @@ yac_expr("Factor(x^2+x-6)")
 yac_expr("D(x) x^2+x-6")
 ```
 
-Using functions easier:
+Using functions easier (using [`magrittr`](https://cran.r-project.org/package=magrittr)'s pipe, `%>%`):
 
 ```{r}
 "x^2+x-6" %>% y_fn("Factor") %>% yac_str()
@@ -82,7 +85,7 @@ Solving equations (removes the `x==` from `yacas`):
 yac_str("Solve(x^2+x-6 == 0, x)")
 yac_solve_str("Solve(x^2+x-6 == 0, x)")
 yac_solve_expr("Solve(x^2+x-6 == 0, x)")
-"x^2+x-6 == 0" %>% y_fn("Solve", "x") %>% yac_solve_str()
+"x^2+x-6 == 0" %>% y_fn("Solve", "x") %>% yac_solve_expr() %>% eval()
 ```
 
 And output in TeX:
@@ -91,10 +94,41 @@ And output in TeX:
 "3/4 + Pi/8" %>% y_fn("Simplify") %>% y_fn("TeXForm") %>% yac_str()
 ```
 
-And arbitrary precision:
+And arbitrary precision (see also the "[Arbitrary-precision arithmetic
+](http://mikldk.github.io/ryacas/articles/arbitrary-precision.html)" vignette):
 
 ```{r}
 yac_str("N(Pi, 50)")
+```
+
+
+### High-level interface
+
+A brief example with a polynomial is:
+```{r}
+x <- yac_symbol("x^2+x-6")
+x
+y_fn(x, "Factor")
+x %>% y_fn("Factor")
+x %>% as_r()
+```
+
+A small matrix example follows:
+
+```{r}
+A <- outer(0:3, 1:4, "-") + diag(2:5)
+a <- 1:4
+B <- yac_symbol(A)
+B
+b <- yac_symbol(a)
+b
+y_fn(B, "Transpose")
+y_fn(B, "Inverse")
+y_fn(B, "Trace")
+B %*% b
+t(B)
+B[, 2:3]
+B %*% solve(B)
 ```
 
 
