### 2019-08-07 #1607 @mikldk
@labarba Thank you very much for considering our submission. We acknowledge your judgement of the submission we made. At the time of submission we had already been working on an extension package for a while that integrated `R` and `yacas` much more closely, e.g. using `yacas` matrices as `R` matrices (with `[` subsetting, finding inverses via `solve()` etc.). The idea was to release it separately afterwards, but based on your judgement we have spent the last week on finishing it and have now fully integrated the functionality into `Ryacas` ([PR#30](https://github.com/mikldk/ryacas/pull/30) and following commit(s)).

May I highlight the updated [README](http://mikldk.github.io/ryacas/index.html) and vignette ["Getting started"](http://mikldk.github.io/ryacas/articles/getting-started.html) as well as the new the vignettes ["The high-level (symbol) interface"](http://mikldk.github.io/ryacas/articles/high-level.html) and ["Arbitrary-precision arithmetic"](http://mikldk.github.io/ryacas/articles/arbitrary-precision.html).

We now believe that the package is much more complete and that it adds much more to `R` than "just" being a thin way to call `yacas` commands. We therefore kindly ask you to reconsider our submission with these additions.
