### 2019-08-21 #1607 @mikldk
Here is a blog post demonstrating usage of the package: <https://mikl.dk/post/2019-pizza-frozen-yogurt/> (mainly under the "Finding the optimal consumption" section).
