### 2019-08-27 03cc46a
Replacing yac_solve_*() with y_rmvars()
