+++ man/y_rmvars.Rd
@@ -1,0 +1,21 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/y-fn.R
+\name{y_rmvars}
+\alias{y_rmvars}
+\title{Remove/strip variable names}
+\usage{
+y_rmvars(x)
+}
+\arguments{
+\item{x}{yacas command}
+}
+\description{
+Remove/strip variable names
+}
+\examples{
+cmd <- "{x == 2, y == 4}"
+yac_str(cmd)
+yac_str(y_rmvars(cmd))
+
+}
+\concept{helper}
