+++ man/yac_solve_expr.Rd
@@ -1,15 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/solve.R
-\name{yac_solve_expr}
-\alias{yac_solve_expr}
-\title{Run yacas Solve() command returning R expression}
-\usage{
-yac_solve_expr(x)
-}
-\arguments{
-\item{x}{yacas command starting with a Solve()}
-}
-\description{
-Run yacas Solve() command returning R expression
-}
-\concept{solve}
