+++ NAMESPACE
@@ -94,13 +94,12 @@ export(upper.tri)
 export(y_fn)
 export(y_ls)
 export(y_print)
+export(y_rmvars)
 export(yac)
 export(yac_assign)
 export(yac_cli)
 export(yac_expr)
 export(yac_silent)
-export(yac_solve_expr)
-export(yac_solve_str)
 export(yac_str)
 export(yac_symbol)
 export(yacas)
