### 2019-08-27 034e8d5
Replacing yac_solve_*() with y_rmvars()
