+++ NAMESPACE
@@ -111,4 +111,5 @@ importFrom(methods,is)
 importFrom(stats,deriv)
 importFrom(utils,loadhistory)
 importFrom(utils,savehistory)
+importFrom(utils,str)
 useDynLib(Ryacas, .registration = TRUE)
