+++ DESCRIPTION
@@ -1,5 +1,5 @@
 Package: Ryacas
-Version: 0.9.0.9122
+Version: 1.0.0
 Title: R Interface to the Yacas Computer Algebra System
 Authors@R: c(
     person(given = "Mikkel Meyer", family = "Andersen",
