+++ inst/exams/poly-factorisation.Rmd
@@ -1,43 +1,0 @@
-```{r data generation, echo = FALSE, results = "hide"}
-possible_answers <- c(1:5) # Must be positive
-answers <- sample(possible_answers, 2, replace = TRUE)
-
-poly <- paste0("(", answers[1L], "*x + ", answers[2L], "*y)^2")
-  
-expr <- poly %>% 
-  y_fn("Expand") %>% 
-  y_fn("Simplify") %>% 
-  y_fn("TeXForm") %>% 
-  yac_str()
-```
-
-Question
-========
-
-Bring 
-$$
-`r expr`
-$$
-on the form 
-$$
-(a \cdot x + b \cdot y)^2
-$$
-for positive $a$ and $b$ and report $a$ and $b$ (in that order).
-
-Solution
-========
-
-$$
-`r expr`
-=
-(`r answers[1L]`x + `r answers[2L]`y)^2
-$$
-
-Thus $a = `r answers[1L]`$ and $b = `r answers[2L]`$.
-
-Meta-information
-================
-exname: Factorisation
-extype: cloze
-exclozetype: string|string
-exsolution: `r paste(answers, collapse = "|")`
