+++ inst/exams/poly-roots.Rmd
@@ -1,36 +1,0 @@
-```{r data generation, echo = FALSE, results = "hide"}
-possible_roots <- c(-(5:1), 1:5)
-roots <- sample(possible_roots, 2, replace = FALSE)
-
-answ <- integer(length(possible_roots))
-answ[match(roots, possible_roots)] <- 1L
-
-poly_tex <- paste0("(x - ", roots[1], ")*(x - ", roots[2], ")") %>% 
-  y_fn("Simplify") %>% 
-  y_fn("TeXForm") %>% 
-  yac_str()
-```
-
-Question
-========
-
-Choose all roots of $`r poly_tex`$:
-
-Answerlist
-----------
-```{r questionlist, echo = FALSE, results = "asis"}
-cat(paste0(paste0("* ", possible_roots), collapse = "\n"))
-```
-
-
-Solution
-========
-
-The roots are `r paste0(sort(roots), collapse = ", ")`.
-
-Meta-information
-================
-exname: Roots in polynomials
-extype: mchoice
-exsolution: `r paste0(answ, collapse = "")`
-exshuffle: FALSE
