### 2019-08-29 399ddf8
Release of CRAN version 1.0.0
