+++ R/y-fn.R
@@ -5,7 +5,11 @@
 #' @param \dots additional arguments to `fn`
 #' 
 #' @examples
+#' y_fn("x^2 - 1", "TeXForm")
+#' yac_str(y_fn("x^2 - 1", "TeXForm"))
+#' 
 #' y_fn("x^2 - 1", "Factor")
+#' yac_str(y_fn("x^2 - 1", "Factor"))
 #' 
 #' cmd <- "x^2 - 1 == 0" %>% y_fn("Solve", "x")
 #' cmd
