### 2019-08-29 824eb2b
minor corrections
