### 2019-09-09 7d303fc
Removed deprecated; added deriv(), Jacobian(), Hessian()
