+++ man/yacas_evaluate.Rd
@@ -1,15 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/legacy.R
-\name{yacas_evaluate}
-\alias{yacas_evaluate}
-\title{Legacy function}
-\usage{
-yacas_evaluate(...)
-}
-\arguments{
-\item{...}{not used}
-}
-\description{
-Legacy function
-}
-\concept{legacy}
