+++ man/Expr.Rd
@@ -1,15 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/legacy.R
-\name{Expr}
-\alias{Expr}
-\title{Legacy function}
-\usage{
-Expr(...)
-}
-\arguments{
-\item{...}{not used}
-}
-\description{
-Legacy function
-}
-\concept{legacy}
