+++ man/I.Rd
@@ -1,15 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/legacy.R
-\name{I}
-\alias{I}
-\title{Legacy function}
-\usage{
-I(...)
-}
-\arguments{
-\item{...}{not used}
-}
-\description{
-Legacy function
-}
-\concept{legacy}
