### 2019-09-18 #1607 @labarba
@openjournals/joss-eics : The author of this rejected submission requests that we have a second look.
