### 2019-09-19 #1607 @kthyng
Hi @drvinceknight! How about another submission? It looks like you are more in the Python community, but as long as you can find reviewers that are comfortable with R, you should be fine.
