### 2019-09-23 #1607 @drvinceknight
@mikldk the pdf fails to compile (see error above), looking at your repo it does not look like there's a `paper.md` file present.
