### 2019-09-23 947a84b
Adding paper again
