+++ paper.bib
@@ -1,0 +1,43 @@
+@manual{R,
+	Address = {Vienna, Austria},
+	Author = {{R Core Team}},
+	Organization = {R Foundation for Statistical Computing},
+	Title = {R: A Language and Environment for Statistical Computing},
+	Url = {https://www.R-project.org/},
+	Year = {2018},
+	Bdsk-Url-1 = {https://www.R-project.org/}}
+
+@TechReport{yacas,
+  author = 	 {Ayal Pinkus and Serge Winnitzky and Grzegorz Mazur},
+  title = 	 {Yacas - Yet another computer algebra system},
+  year = 	 {2016},
+  Url = {https://yacas.readthedocs.io/en/latest/},
+}
+
+@Article{Rcpp,
+  title = {{Extending \textit{R} with \textit{C++}: A Brief Introduction to \textit{Rcpp}}},
+  author = {D Eddelbuettel and J J Balamuta},
+  journal = {PeerJ Preprints},
+  year = {2017},
+  month = {aug},
+  volume = {5},
+  pages = {e3188v1},
+  issn = {2167-9843},
+  url = {https://doi.org/10.7287/peerj.preprints.3188v1},
+  doi = {10.7287/peerj.preprints.3188v1},
+}
+
+@inproceedings{Pinkus2002,
+ author = {Pinkus, Ayal Z. and Winitzki, Serge},
+ title = {{YACAS: A Do-It-Yourself Symbolic Algebra Environment}},
+ booktitle = {Proceedings of the Joint International Conferences on Artificial Intelligence, Automated Reasoning, and Symbolic Computation},
+ series = {AISC '02/Calculemus '02},
+ year = {2002},
+ isbn = {3-540-43865-3},
+ pages = {332--336},
+ numpages = {5},
+ url = {http://doi.org/10.1007/3-540-45470-5_29},
+ doi = {10.1007/3-540-45470-5_29},
+ publisher = {Springer-Verlag},
+ address = {London, UK, UK},
+} 
