### 2019-09-23 #1607 @arfon
> @arfon, sorry for the ping but this paper does not appear in my dashboard: have I done something incorrectly?

Should be showing up now. We just found an edge-case (bug) in the system :-)
