### 2019-09-24 #1607 @drvinceknight
Hi @matloff, @wrathematics, @ajank and @acolum would you be interested in reviewing this paper for JOSS? It is presenting a piece of software that interfaces to the `yacas` computer algebra system.

If you have not reviewed for JOSS before here is some documentation describing the process: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html and I'd be very happy to answer any questions you have.

Once two reviewers have agreed I will open a separate review issue where the review process can take place.
