### 2019-09-24 7674472
Adding y_rmvars and [[ for yac_symbol; minor documentation typos fixed
