+++ man/sub-sub-.yac_symbol.Rd
@@ -1,0 +1,18 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/yac-symbol.R
+\name{[[.yac_symbol}
+\alias{[[.yac_symbol}
+\title{Extract or replace parts of an object}
+\usage{
+\method{[[}{yac_symbol}(x, i)
+}
+\arguments{
+\item{x}{A \code{yac_symbol}.}
+
+\item{i}{row indices specifying elements to extract or replace}
+
+\item{j}{column indices specifying elements to extract or replace}
+}
+\description{
+Extract or replace parts of an object
+}
