+++ NAMESPACE
@@ -4,6 +4,7 @@ S3method("%*%",default)
 S3method("%*%",yac_symbol)
 S3method("[",yac_symbol)
 S3method("[<-",yac_symbol)
+S3method("[[",yac_symbol)
 S3method("diag<-",default)
 S3method("diag<-",yac_symbol)
 S3method(Hessian,yac_symbol)
@@ -29,6 +30,8 @@ S3method(upper.tri,default)
 S3method(upper.tri,yac_symbol)
 S3method(y_fn,default)
 S3method(y_fn,yac_symbol)
+S3method(y_rmvars,default)
+S3method(y_rmvars,yac_symbol)
 S3method(yac,default)
 S3method(yac,yac_symbol)
 S3method(yac_assign,default)
