### 2019-09-25 #1763 @whedon
[REVIEW]: Ryacas: A computer algebra system in R
**Submitting author:** @mikldk (<a href="http://orcid.org/0000-0002-0234-0266">Mikkel Meyer Andersen</a>)
**Repository:** <a href="https://github.com/mikldk/ryacas" target ="_blank">https://github.com/mikldk/ryacas</a>
**Version:** v1.1.0
