### 2019-09-30 mikldk/ryacas#32 @MikeLydeamore
JOSS Review Issues
Thank you for your work on this software package and corresponding paper. CAS is somewhat out of my field of expertise (other than fairly basic use) but should the need arise I think this would make a useful addition to my workflow.

From a usability perspective, everything worked as described in the repository and the few tests I ran also went successfully. There are a couple of minor issues that I have identified:

- The repository is missing an OSI compatible license.
- The vignettes do not appear to be accessible from within R (and weren't listed by using `vigentte(pkg="ryacas")`. It's unclear to be whether they are supposed to be, but I think it would be a nice addition and in line with R package standards.
- There are currently no guidelines for contribution, reporting issues or seeking support.

Otherwise, I believe the package/paper is ready for publication.
