### 2019-09-30 #1763 @MikeLydeamore
Hi @drvinceknight, I have completed my review and have recommend minor revisions. I've created an issue in the source repository that lists these: https://github.com/mikldk/ryacas/issues/32
