### 2019-09-30 777cf33
Add info about contributing, installing with vignettes and license addressing #32
