+++ README.Rmd
@@ -35,11 +35,19 @@ pretty printing and translating R to TeX.
 
 ## Install from GitHub ##
 
-To build and install from github using R 3.3.0 (or later) and the R `devtools` package 1.11.0 (or later) run this command from within R:
+To build and install from Github using R 3.3.0 (or later) and the R `devtools` package 1.11.0 (or later) run this command from within `R`:
+
+```
+devtools::install_github("mikldk/ryacas0", 
+                         build_opts = c("--no-resave-data", "--no-manual"))
+```
+
+You can also install the package without vignettes if needed as follows:
 
 ```
 devtools::install_github("mikldk/ryacas")
 ```
+
 
 ## Online info ##
 
@@ -47,6 +55,12 @@ For vignettes, overview, pointers to additional information, installation
 instructions and a sample session see <http://mikldk.github.io/ryacas/>. 
 
 Yacas documentation can be found at http://yacas.readthedocs.org/.
+
+## Contribute, issues, and support ##
+
+Please use the issue tracker at <https://github.com/mikldk/ryacas/issues> 
+if you want to notify us of an issue or need support.
+If you want to contribute, please either create an issue or make a pull request.
 
 ## Brief examples ##
 
