### 2019-09-30 mikldk/ryacas#32 @mikldk
Thanks for your review and for helping making the package better by catching these shortages. I hope that commit 777cf33459d9b8349b1863149182c38652b4b23b have addressed your points.
