### 2019-09-30 #1763 @EduPH
> **Contribution and authorship:** Has the submitting author (@mikldk) made major contributions to the software? Does the full list of paper authors seem appropriate and complete?

@mikldk, I can see that there are more authors cited at the README of the repository. Should they be added?

> **Example usage:** Do the authors include examples of how to use the software (ideally to solve real-world analysis problems).

I see a lot of examples of the use of Ryacas in the documentation. However, I think they could be described a little bit more. Some of them have no description at all (http://mikldk.github.io/ryacas/articles/low-level.html#sums)

> **Functionality documentation:** Is the core functionality of the software documented to a satisfactory level (e.g., API method documentation)?
The same as the previous point. I understand that is based on yacas, so you can use yacas documentation as solid ground. But some parts of documentation are too many lines of codes without explanations.

Typo in the paper:
Hoever -> However

--
However, I think that the documentation is useful, and your work is great and useful. Apart from this, I have no other objections to the publication.
