+++ man/isolate.Rd
@@ -1,0 +1,19 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/yac-symbol.R
+\name{isolate}
+\alias{isolate}
+\title{Isolate variable in a yac symbol}
+\usage{
+isolate(x, var)
+}
+\arguments{
+\item{x}{A \code{yac_symbol}}
+
+\item{var}{variable to find root with respect to}
+}
+\description{
+Correspond to solve, but called something
+differently to avoid confusion.
+}
+\concept{yac_symbol}
+\concept{yac_symbol_solve}
