+++ NAMESPACE
@@ -11,12 +11,15 @@ S3method(Hessian,yac_symbol)
 S3method(Jacobian,yac_symbol)
 S3method(Math,yac_symbol)
 S3method(Ops,yac_symbol)
+S3method(as.character,yac_symbol)
 S3method(as_r,default)
 S3method(as_r,yac_symbol)
 S3method(deriv,yac_symbol)
 S3method(diag,default)
 S3method(diag,yac_symbol)
 S3method(dim,yac_symbol)
+S3method(equal_to,yac_symbol)
+S3method(isolate,yac_symbol)
 S3method(length,yac_symbol)
 S3method(lower.tri,default)
 S3method(lower.tri,yac_symbol)
@@ -50,7 +53,9 @@ export(Jacobian)
 export(as_r)
 export(as_y)
 export(diag)
+export(equal_to)
 export(get_output_width)
+export(isolate)
 export(lower.tri)
 export(set_output_width)
 export(simplify)
