### 2019-10-01 0f60bcf
Extending solve()
