+++ R/yac-symbol.R
@@ -358,35 +358,115 @@ t.yac_symbol <- function(x) {
   return(v)
 }
 
-#' Solve a system of equations
-#' 
-#' This generic function solves the 
-#' equation $a %*% x = b$ for $x$.
-#' 
-#' @param a If `yac_symbol` treat as such, else 
-#' call [base::solve()].
-#' @param b If `yac_symbol` treat as such, else 
-#' call [base::solve()].
-#' @param \dots Not used
-#' 
-#' @export
-solve.yac_symbol <- function(a, b, ...) {
-  # FIXME: Currently Inverse, but implement general case
-  if (!missing(b)) {
-    stop("solve() currently only supports finding inverse")
-  }
+solve_yac_symbol_matrixinverse <- function(a) {
+  stopifnot(a$is_mat)
   
-  # To see if x is indeed a matrix
-  y_res <- yac_str(a$yacas_cmd)
-  y <- yac_symbol(y_res)
-
-  stopifnot(y$is_mat)
-  
-  z <- y_fn(x = y, fn = "Inverse")
+  z <- y_fn(x = a, fn = "Inverse")
   z_res <- yac_str(z)
   v <- yac_symbol(z_res)
   
   return(v)
+}
+
+solve_yac_symbol_linearsolve <- function(a, b) {
+  stopifnot(a$is_mat)
+  stopifnot(b$is_vec)
+  
+  stopifnot(nrow(a) == length(b))
+  
+  cmd <- paste0("SolveMatrix(", a$yacas_cmd, ", ", b$yacas_cmd, ")")
+  v <- yac_symbol(cmd)
+  
+  return(v)
+}
+  
+#' Solve a system of equations
+#' 
+#' This generic function solves the 
+#' equation $a x = b$ for $x$.
+#' 
+#' When `a` is a matrix and `b` not provided,
+#' this finds the inverse of `a`.
+#' When `a` is a matrix and a vector `b` is provided, the 
+#' linear system of equations is solved.
+#' 
+#' Note that solving non-linear equations:
+#' 
+#' * `solve(a, b)`: find roots of `a` for variable `b`, i.e. yacas `Solve(a == 0, b)`
+#' * `solve(a, b, v)`: find solutions to `a == b` for variable `v`, i.e. yacas `Solve(a == b, v)`
+#' 
+#' @param a A `yac_symbol` 
+#' @param b A `yac_symbol` or a value, see details and examples.
+#' @param \dots See details and examples.
+#' 
+#' @examples 
+#' A <- outer(0:3, 1:4, "-") + diag(2:5)
+#' a <- 1:4
+#' B <- yac_symbol(A)
+#' b <- yac_symbol(a)
+#' solve(A)
+#' solve(B)
+#' solve(A, a)
+#' solve(B, b)
+#' 
+#' poly <- yac_symbol("x^2 - x - 6")
+#' solve(poly, "x")    # Solve(poly == 0, x)
+#' solve(poly, 3, "x") # Solve(poly == 3, x)
+#' 
+#' @export
+solve.yac_symbol <- function(a, b, ...) {
+  if (!is(a, "yac_symbol")) {
+    stop("'a' must be a yac_symbol")
+  }
+
+  # If only a given:
+  if (missing(b)) {
+    # If a matrix:
+    if (a$is_mat) {
+      return(solve_yac_symbol_matrixinverse(a))
+    } else if (a$is_vec) {
+      stop("a cannot be a yac_symbol vector")
+    } else if (!a$is_vec) {
+      # a not matrix nor vector
+      stop("Trying to find roots? Please provide variable name, too")
+    } else {
+      stop("Should not occur")
+    }
+  } else {
+    # b also given:
+    
[…]
