### 2019-10-01 7190e22
Minor typos
