### 2019-10-01 #1763 @mikldk
@EduPH, thanks for good comments! I have tried making a few corrections in documentation ([diff](https://github.com/mikldk/ryacas/commit/c8e4af65bd4d57c9c907840d78daed37e3393c51#diff-c87dadc225fad3d4b5c2429f3261b341)) and in the paper ([diff](https://github.com/mikldk/ryacas/compare/459fe7f200b044cae5fa3bb0ff8589cf42773653..7190e22b4019e0127c6f980708450aa8e45aa5b1)) as suggested. I have also updated `solve()` so that it works in general now ([docs](http://mikldk.github.io/ryacas/articles/high-level.html#solving-equations)). The author list is accurate.
