### 2019-10-01 6ee23d3
Internal function to easier generate code when reporting yacas issues
