### 2019-10-05 #1763 @mikldk
@drvinceknight I see that I wrote version 1.0.0, but it actually should be version 1.1.0. Is that something that you can change? Sorry about this.
