### 2019-10-08 d602873
Preparing v1.1.0 release
