### 2019-10-08 #1763 @mikldk
@drvinceknight: I have made a [release of v1.1.0](https://github.com/mikldk/ryacas/releases/tag/1.1.0) (via [a small commit](https://github.com/mikldk/ryacas/commit/d60287363844ef75f4dc5e339abc7bc6e588c14a)). I have archived this on Zenodo: <https://doi.org/10.5281/zenodo.3475813>. Let me know if I need to do more.

@EduPH and @MikeLydeamore: Thanks a lot for your help with a fast and thorough review.
