### 2019-10-08 #1763 @mikldk
@drvinceknight: I have made a release of v1.1.0. I have archived this on Zenodo: https://zenodo.org/record/3476220 (DOI: [10.5281/zenodo.3476220](https://doi.org/10.5281/zenodo.3476220)). Let me know if I need to do more.

I also updated the package at CRAN: <https://CRAN.R-project.org/package=Ryacas>.

@EduPH and @MikeLydeamore: Thanks a lot for your help with a fast and thorough review.
