### 2019-10-09 #1763 @drvinceknight
@openjournals/joss-eics this paper is ready to be accepted
