### 2019-10-09 #1763 @labarba
Please edit the Zenodo archive's metadata so the title and author list matches the JOSS paper.
