### 2019-10-09 #1763 @drvinceknight
> The software archive is still pending:

I had checked that @labarba (based on the checklist you circulated on a previous review) but please let me know if I could have made that clearer. (Should I have asked Whedon to archive as you did?).

> Please edit the Zenodo archive's metadata so the title and author list matches the JOSS paper

I see that I missed that. Apologies.
