### 2019-10-09 #1763 @labarba
> Should I have asked Whedon to archive as you did?

Yes, the handling editor runs the command `@whedon set <doi> as archive` before handing it over to AEiCs.
