### 2019-10-09 #1763 @mikldk
> Please edit the Zenodo archive's metadata so the title and author list matches the JOSS paper.

@labarba @drvinceknight : This has now been done: <https://zenodo.org/record/3476220>.
