### 2019-10-09 #1763 @drvinceknight
> Yes, the handling editor runs the command `@whedon set <doi> as archive` before handing it over to AEiCs.

Thanks, I also see where I missed that in the docs. Apologies for the trouble :+1:
