### 2019-10-09 #1763 @labarba
Congratulations, @mikldk, your JOSS paper is now published! 🚀

Huge thanks to our editor: @drvinceknight, and the reviewers: @MikeLydeamore, @EduPH — your contribution to JOSS is greatly appreciated!
