### 2019-10-09 #1763 @mikldk
I would repeat after @whedon: Huge thanks to our editor @drvinceknight and the reviewers @MikeLydeamore and @EduPH -- your effort is greatly appreciated!
