### 2019-10-09 #1763 @drvinceknight
A personal thank you to the editors from me: this doesn't happen without you. Congratulations @mikldk and a further personal thanks to @labarba for your patience and guidance :+1:
